backend/pyproject.toml
[tool.poetry]
name = "counter_app"
version = "0.1.0"
description = ""
authors = ["liron.lavy@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.25"}
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
async-exit-stack = "^1.0.1"
async_generator = "^1.10"
pyhumps = "^3.0.2"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"
isort = "^5.9.3"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"

[[tool.mypy.overrides]]
module = [
    "humps",
    "uvicorn"
]
ignore_missing_imports = true
